+++ apps/ascent/components/docs/index.ts
@@ -1,2 +1,3 @@
 export { default as DocsPage } from './DocsPage'
 export { default as Sidebar } from './Sidebar'
+export { SidebarItem } from './SidebarItem'
