+++ apps/ascent/components/features/Documentation/Preview.tsx
@@ -1,4 +1,3 @@
-'use client'
 import React from 'react'
 
 const Preview = ({ children }: { children: React.ReactNode }) => {
